move map controls to map elements
